backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
gunicorn==21.2.0
facenet-pytorch==2.5.3
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.2
Pillow==10.1.0
requests==2.31.0
waitress==2.1.2
